refactor: P3 TK companion 抽取（功能不变）
外提 ops alert dimensions、gateway_forward Cloudwise/passthrough/ToolSearch、
billing fallback 表、scheduling window recovery、buffered SSE 脚手架、
runtime fastpath、setting signup/pricing 与 Responses compact 分支到
*_tk_*.go，上游形文件只留薄调用点。同步 gateway-tk / pricing-availability
sentinel。
no-web-impact
sentinel-registry-reviewed
